refactor sidecar RuntimeInfo from interface.rs to separate file
