+++ sidecar/src/service/mod.rs
@@ -4,6 +4,7 @@
 pub use instance_id::InstanceId;
 pub use queue_id::QueueId;
 pub use request_identification::{RequestIdentification, RequestIdentifier};
+pub use runtime_info::RuntimeInfo;
 pub use runtime_metadata::RuntimeMetadata;
 pub use serialized_tracer_header_tags::SerializedTracerHeaderTags;
 pub use session_info::SessionInfo;
@@ -11,9 +12,11 @@ pub use sidecar_interface::{
     SidecarInterface, SidecarInterfaceClient, SidecarInterfaceRequest, SidecarInterfaceResponse,
 };
 pub use sidecar_server::SidecarServer;
+
 mod instance_id;
 pub mod queue_id;
 mod request_identification;
+mod runtime_info;
 mod runtime_metadata;
 mod serialized_tracer_header_tags;
 mod session_info;
